First assingment commit, issues with Intellji meant I was unable to connect to the resposiatory for a while.
This commit has my Random_Forest.ipynb which saw lots of trial and error, mostly due to computational power and one-hot encoding but I was able to fix it by using fewer trees, going less deep and less splits. This reduced the computation time where it was getting "unlucky" and spending hours to an average of 12.5 minutes. I also found it did not drastically change the precision and recall rates